requirements.txt
mysql-connector-python==8.3.0
Faker==24.4.0
redis==5.0.3
kafka-python==2.2.10
python-dotenv==1.0.1